Changes to be committed: 	deleted:    combinedFile_1.ts 	modified:   src/app/api/config/porcentagens/route.js 	modified:   src/app/api/historico-cliente/create/route.js 	new file:   src/app/api/historico-cliente/list/route.js 	modified:   src/app/api/jogos/create/route.js 	modified:   src/app/api/jogos/list/route.js 	modified:   src/app/api/resultados/create/route.js 	modified:   src/app/components/dashboard/Admin/AdminDashboard.jsx 	modified:   src/app/components/dashboard/Admin/Configuracoes.jsx 	modified:   src/app/components/dashboard/Admin/Financeiro.jsx 	modified:   src/app/components/dashboard/Admin/GameManagement.jsx 	modified:   src/app/components/dashboard/Admin/PorcentagensConfig.jsx 	modified:   src/app/components/dashboard/Admin/ResultadosManagement.jsx 	modified:   src/app/components/dashboard/Sidebar.jsx 	new file:   src/app/utils/updateGameStatuses.js 	modified:   teste.js

+++ src/app/components/dashboard/Admin/GameManagement.jsx
@@ -55,6 +55,18 @@ const GameManagement = () => {
 
     try {
       const token = localStorage.getItem('token');
+      if (!token) {
+        toast({
+          title: 'Token não encontrado.',
+          description: 'Por favor, faça login novamente.',
+          status: 'warning',
+          duration: 5000,
+          isClosable: true,
+        });
+        setLoading(false);
+        return;
+      }
+
       const response = await axios.get('/api/jogos/list', {
         headers: {
           Authorization: `Bearer ${token}`,
@@ -94,6 +106,17 @@ const GameManagement = () => {
     try {
       const updatedVisibility = !jogo.visibleInConcursos;
       const token = localStorage.getItem('token');
+      if (!token) {
+        toast({
+          title: 'Token não encontrado.',
+          description: 'Por favor, faça login novamente.',
+          status: 'warning',
+          duration: 5000,
+          isClosable: true,
+        });
+        return;
+      }
+
       await axios.put(`/api/jogos/${jogo.slug}`, { visibleInConcursos: updatedVisibility }, {
         headers: {
           Authorization: `Bearer ${token}`,
@@ -124,6 +147,17 @@ const GameManagement = () => {
 
     try {
       const token = localStorage.getItem('token');
+      if (!token) {
+        toast({
+          title: 'Token não encontrado.',
+          description: 'Por favor, faça login novamente.',
+          status: 'warning',
+          duration: 5000,
+          isClosable: true,
+        });
+        return;
+      }
+
       await axios.delete(`/api/jogos/${jogo.slug}`, {
         headers: {
           Authorization: `Bearer ${token}`,
